+++ ide-gen-proto.sh
@@ -73,7 +73,7 @@ gen_proto_go checkoutservice
 # gen_proto_ruby email
 gen_proto_ts frontend
 gen_proto_ts react-native-app
-gen_proto_js paymentservice
+gen_proto_js payment
 gen_proto_go productcatalogservice
 # gen_proto_php quote
 gen_proto_python recommendationservice